Add a method to search for connected joysticks on Android
- Fix a bug on Android related to counter ticks in the timer
- Add a constructor to the Android input manager class
- Make the method to init the accelerometer sensor private
- Fix a warning in the SDL input manager

+++ src/ncTimer.cpp
@@ -96,7 +96,7 @@ unsigned long long int ncTimer::Counter()
 	{
 		struct timespec now;
 		clock_gettime(CLOCK_MONOTONIC, &now);
-		ullCounter = now.tv_sec * s_ulFrequency + now.tv_nsec;
+		ullCounter = (unsigned long long int)now.tv_sec * s_ulFrequency + (unsigned long long int)now.tv_nsec;
 	}
 	else
 	{
